+++ ThePartingOfSarah/Space.h
@@ -21,17 +21,16 @@ public:
 	void removeStaticActor(Actor* actor);
 	void removeLowStaticActor(Actor* actor);
 
-
 private:
 	void updateMovementUp(Actor* dynamicActor, bool flying);
 	void updateMovementLeft(Actor* dynamicActor, bool flying);
 	void updateMovementRight(Actor* dynamicActor, bool flying);
 	void updateMovementBottom(Actor* dynamicActor, bool flying);
 
-	void checkCollisionUp(Actor* dynamicActor, Actor* staticActor, int* possibleMovement);
-	void checkCollisionLeft(Actor* dynamicActor, Actor* staticActor, int* possibleMovement);
-	void checkCollisionRight(Actor* dynamicActor, Actor* staticActor, int* possibleMovement);
-	void checkCollisionDown(Actor* dynamicActor, Actor* staticActor, int* possibleMovement);
+	void checkCollisionUp(Actor* dynamicActor, Actor* staticActor, float* possibleMovement);
+	void checkCollisionLeft(Actor* dynamicActor, Actor* staticActor, float* possibleMovement);
+	void checkCollisionRight(Actor* dynamicActor, Actor* staticActor, float* possibleMovement);
+	void checkCollisionDown(Actor* dynamicActor, Actor* staticActor, float* possibleMovement);
 
 	list<Actor*> dynamicActors;
 	list<Actor*> flyingDynamicActors;
